Moved marker classes to they own package
