+++ collect_app/src/test/java/org/odk/collect/android/widgets/support/NoOpMapFragment.kt
@@ -3,8 +3,8 @@ package org.odk.collect.android.widgets.support
 import androidx.fragment.app.Fragment
 import org.odk.collect.maps.MapFragment
 import org.odk.collect.maps.MapPoint
-import org.odk.collect.maps.MarkerDescription
-import org.odk.collect.maps.MarkerIconDescription
+import org.odk.collect.maps.markers.MarkerDescription
+import org.odk.collect.maps.markers.MarkerIconDescription
 
 class NoOpMapFragment : Fragment(), MapFragment {
