+++ maps/src/main/java/org/odk/collect/maps/markers/MarkerDescription.kt
@@ -1,4 +1,7 @@
-package org.odk.collect.maps
+package org.odk.collect.maps.markers
+
+import org.odk.collect.maps.MapFragment
+import org.odk.collect.maps.MapPoint
 
 data class MarkerDescription(
     val point: MapPoint,
