+++ maps/src/main/java/org/odk/collect/maps/markers/MarkerIconCreator.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.maps
+package org.odk.collect.maps.markers
 
 import android.content.Context
 import android.graphics.Bitmap
