+++ osmdroid/src/main/java/org/odk/collect/osmdroid/OsmDroidMapFragment.java
@@ -47,9 +47,9 @@ import org.odk.collect.maps.MapConfigurator;
 import org.odk.collect.maps.MapFragment;
 import org.odk.collect.maps.MapFragmentDelegate;
 import org.odk.collect.maps.MapPoint;
-import org.odk.collect.maps.MarkerDescription;
-import org.odk.collect.maps.MarkerIconCreator;
-import org.odk.collect.maps.MarkerIconDescription;
+import org.odk.collect.maps.markers.MarkerDescription;
+import org.odk.collect.maps.markers.MarkerIconCreator;
+import org.odk.collect.maps.markers.MarkerIconDescription;
 import org.odk.collect.maps.layers.MapFragmentReferenceLayerUtils;
 import org.odk.collect.maps.layers.ReferenceLayerRepository;
 import org.odk.collect.settings.SettingsProvider;
